+++ packages/components/src/Form/Fields/FieldBase.tsx
@@ -23,37 +23,43 @@
  SOFTWARE.
 
  */
-
-import React, { forwardRef, Ref } from 'react'
+import React from 'react'
 import styled from 'styled-components'
-import { v4 as uuid } from 'uuid'
-import { useFormContext } from '../../Form'
-import { Select, SelectProps } from '../../Inputs/Select/Select'
-import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
+import { FontSizes, FontWeights } from '@looker/design-tokens'
+import { ValidationMessageProps } from '../ValidationMessage/ValidationMessage'
 
-export interface FieldSelectProps extends FieldProps, SelectProps {}
+export interface FieldBaseProps {
+  className?: string
+  disabled?: boolean
+  /**
+   * Defines the label for the field.
+   */
+  label?: string
+  /**
+   * Specifies the fontWeight of the internal Label.
+   * TODO - Deprecate usage in HT, then here.
+   */
+  labelFontSize?: FontSizes
+  /**
+   * Specifies the fontWeight of the internal Label.
+   */
+  labelFontWeight?: FontWeights
+  /**
+   * Whether or not the field should display a `*` denoting it is required.
+   */
+  required?: boolean
+  /**
+   *
+   * Holds the type of validation (error, warning, etc.) and corresponding message.
+   */
+  validationMessage?: ValidationMessageProps
+}
 
-const FieldSelectComponent = forwardRef(
-  (props: FieldSelectProps, ref: Ref<HTMLInputElement>) => {
-    const validationMessage = useFormContext(props)
-    const { id = uuid() } = props
-    return (
-      <Field
-        id={id}
-        validationMessage={validationMessage}
-        {...pickFieldProps(props)}
-      >
-        <Select
-          {...omitFieldProps(props)}
-          id={id}
-          validationType={validationMessage && validationMessage.type}
-          ref={ref}
-        />
-      </Field>
-    )
-  }
-)
-
-FieldSelectComponent.displayName = 'FieldSelectComponent'
-
-export const FieldSelect = styled(FieldSelectComponent)``
+export const RequiredStar = styled((props) => (
+  <span {...props} aria-hidden="true">
+    {' '}
+    *
+  </span>
+))`
+  color: ${(props) => props.theme.colors.palette.red500};
+`