+++ packages/components/src/Form/ValidationMessage/ValidationMessage.tsx
@@ -24,31 +24,14 @@
 
  */
 
-import React, { forwardRef, Ref } from 'react'
-import {
-  CompatibleHTMLProps,
-  CustomizableAttributes,
-  layout,
-  LayoutProps,
-  reset,
-  space,
-  SpaceProps,
-  typography,
-  TypographyProps,
-} from '@looker/design-tokens'
+import React, { FC } from 'react'
+import { reset } from '@looker/design-tokens'
 import styled from 'styled-components'
 
 export type ValidationType = 'error'
 
-export const CustomizableValidationMessageAttributes: CustomizableAttributes = {
-  fontSize: 'xsmall',
-}
-
-export interface ValidationMessageProps
-  extends LayoutProps,
-    SpaceProps,
-    TypographyProps,
-    CompatibleHTMLProps<HTMLDivElement> {
+export interface ValidationMessageProps {
+  className?: string
   /**
    * The type of validation, therefore changing the message's text color. Accepts: error.
    */
@@ -59,28 +42,18 @@ export interface ValidationMessageProps
   message?: string
 }
 
-const ValidationMessageBase = styled.div.attrs(
-  (props: ValidationMessageProps) => ({
-    fontSize:
-      props.fontSize || CustomizableValidationMessageAttributes.fontSize,
-  })
-)<ValidationMessageProps>`
-  ${reset}
-  ${(props) =>
-    props.type === 'error'
-      ? `color: ${props.theme.colors.palette.red500};`
-      : ''}
-  ${layout}
-  ${space}
-  ${typography}
-`
+const ValidationMessageLayout: FC<ValidationMessageProps> = ({
+  className,
+  message,
+}) => <div className={className}>{message}</div>
 
-export const ValidationMessage = forwardRef(
-  ({ message, ...props }: ValidationMessageProps, ref: Ref<HTMLDivElement>) => (
-    <ValidationMessageBase mr="xsmall" mt="xsmall" {...props} ref={ref}>
-      {message}
-    </ValidationMessageBase>
-  )
-)
+export const ValidationMessage = styled(ValidationMessageLayout)`
+  ${reset}
+
+  font-size: ${(props) => props.theme.fontSizes.xsmall};
+
+  ${({ theme, type }) =>
+    type === 'error' && `color: ${theme.colors.palette.red500};`}
+`
 
 ValidationMessage.displayName = 'ValidationMessage'